+++ .ladle/components.tsx
@@ -1,6 +1,5 @@
 import type { GlobalProvider } from "@ladle/react";
-import React, { useEffect } from "react";
-import { ThemeProvider } from "../src/components/ThemeProvider";
+import { useEffect } from "react";
 import "./theme.css";
 
 export const Provider: GlobalProvider = ({ children, globalState }) => {
@@ -11,5 +10,5 @@ export const Provider: GlobalProvider = ({ children, globalState }) => {
       document.documentElement.classList.remove("vkui:dark");
     }
   }, [globalState.theme]);
-  return <ThemeProvider>{children}</ThemeProvider>;
+  return children;
 };